airflow/requirements.txt
altair==4.2.2
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.5
boto==2.49.0
boto3==1.26.74
botocore==1.29.74
branca==0.6.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-default-group-wheel==1.2.2
contourpy==1.0.7
crypto==1.4.1
cryptography==39.0.1
cycler==0.11.0
decorator==5.1.1
diagrams==0.23.3
ecdsa==0.18.0
entrypoints==0.4
exceptiongroup==1.1.0
extra-streamlit-components==0.1.56
Faker==17.0.0
fastapi==0.92.0
favicon==0.7.0
folium==0.14.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
htbuilder==0.6.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.2
jmespath==1.0.1
jose==1.0.0
jsonschema==4.17.3
jwt==1.3.1
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
markdownlit==0.0.5
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
more-itertools==9.0.0
Naked==0.1.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
pymdown-extensions==9.9.2
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
semver==2.13.0
shellescape==3.8.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
sqlite-fts4==1.0.3
sqlite-utils==3.30
st-annotated-text==3.0.0
starlette==0.25.0
streamlit==1.18.1
streamlit-authenticator==0.2.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.6
streamlit-faker==0.0.2
streamlit-folium==0.11.1
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typed-ast==1.5.4
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
watchdog==2.2.1
zipp==3.14.0
